requirements.txt
# Core dependencies
chromadb>=0.4.0
sentence-transformers>=2.2.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0

# UI
streamlit>=1.28.0

# Testing
pytest>=7.0.0
